backend/requirements.txt
fastapi
uvicorn
ultralytics
python-multipart
groq
python-dotenv
opencv-python-headless
numpy
